Move files to service
