+++ src/main/java/service/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package main;
+package service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
