+++ src/main/java/service/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package main;
+package service;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
